CTPy-10 Better calculation of stationarity for starting data collection
Use Ewens and Gillespie (1974) approximation to calculate time to stationarity for WF-IA process.  This is conservative because I'm using the diploid value in a haploid process.  The population evolves for this time period, and then we begin data collection for the configured simulation sample length.  These values are not yet saved in the simulation data object, but work properly in both sweep and interactive mode.

Also, sphinx documentation structure is now working.