CTPy-12 Prototype of external classification script and data model:
Script has functions for calculating and caching dimension/mode boundaries based upon maxalleles and a number of modes to chop each dimension into.

Currently, the dimension is chopped into equal partitions, and the number is hardwired, pending a data model for describing a classification.  But it works, and post processing is fast.