CTPy-22 Analytics script to merge primary trait richness, evenness, into post-classification data
Created script and modules to calculate trait level richness and evenness stats from the fulldataset -- the richness and other tables in the "raw" database are deprecated since they only apply to the simulation run itself, not the subsampled full data set.  The run-experiment-sequence.sh script is updated, as is the data export script.